+++ ParkingSystem/Entities/Parking.cs
@@ -7,6 +7,6 @@
         public int NumberOfPlaces { get; set; }
         public TimeSpan OpeningTime { get; set; }
         public TimeSpan ClosingTime { get; set; }
-        public int PricePerHour { get; set; }
+        public decimal PricePerHour { get; set; }
     }
 }
